replace signatures, keys and signing with ethers-core and k256
