+++ Cargo.toml
@@ -15,7 +15,7 @@ path = "src/server.rs"
 
 [[bin]]
 name = "cli"
-path = "src/cli.rs"
+path = "src/cli/cli.rs"
 
 [[bin]]
 name = "client"
